gnome-classic: Define a gnome-session session
For one, this is simply more correct: the gnome-classic session is a
different session from the default gnome session. So let's define it as
such

For two, this enables us to directly tell gnome-shell which shell mode
to use, rather than passing it through an environment variable. In other
words, this enables us to integrate with
https://gitlab.gnome.org/GNOME/gnome-shell/-/merge_requests/3960